Move test code from Registry to RegistrySpec
